fix two index types not compatible with x86

+++ source/url.d
@@ -1180,7 +1180,7 @@ string punyDecode(string input) {
 	// let i = 0
 	// let bias = initial_bias
 	// let output = an empty string indexed from 0
-	ulong i = 0;
+	size_t i = 0;
 	auto bias = initialBias;
 	dchar[] output;
 	// This reserves a bit more than necessary, but it should be more efficient overall than just
@@ -1200,7 +1200,7 @@ string punyDecode(string input) {
 	}
 
  	// while the input is not exhausted do begin
-	ulong pos = 0;
+	size_t pos = 0;
 	while (pos < input.length) {
  	//   let oldi = i
  	//   let w = 1
